move video interface fetch to storage core
